Move the package table into its own folder
`src/components/package-table/` holds the Astro wrapper, the Svelte
island, and a new `data.ts` with the pure logic: the document format,
the positional unpacking, version comparison, the per-mutex row
derivation, filters and sorters. The component keeps the reactive
state, the template and the styles.
