+++ src/components/package-table/PackageTable.astro
@@ -1,6 +1,6 @@
 ---
-import type { Distro } from "../data/distros";
-import { browseUrl } from "../data/distros";
+import type { Distro } from "../../data/distros";
+import { browseUrl } from "../../data/distros";
 import PackageTable from "./PackageTable.svelte";
 
 interface Props {
