+++ src/pages/[distro].astro
@@ -1,7 +1,7 @@
 ---
 import StarlightPage from "@astrojs/starlight/components/StarlightPage.astro";
 import { Aside, Code } from "@astrojs/starlight/components";
-import PackageTable from "../components/PackageTable.astro";
+import PackageTable from "../components/package-table/PackageTable.astro";
 import type { Distro } from "../data/distros";
 import { browseUrl, DISTROS, monthYear, title } from "../data/distros";
